+++ client/src/scripts/objects/parachute.ts
@@ -2,7 +2,7 @@ import { Sprite2D, ABParticle2D } from "common/engine/web.ts"
 import { GameObject } from "../others/gameObject.ts"
 import { GameObjectType, zIndexes } from "common/scripts/others/constants.ts"
 import { FloorKind, Floors, FloorType } from "common/scripts/others/terrain.ts";
-import { CircleHitbox2D, ColorM, random, Stream, v2, v2m } from "common/engine/core.ts";
+import { CircleHitbox2D, ColorM, matrix4, Numeric, random, Stream, v2, v2m } from "common/engine/core.ts";
 
 export class Parachute extends GameObject{
     ////////////////////////////
@@ -80,10 +80,14 @@ export class Parachute extends GameObject{
             this.on_landed()
             this.destroy()
         }
-        const s=v2(1.2,1.2)
-        v2m.scale(s,s,1-this.time/this.parachute_data.lifetime)
-        v2m.add(this.sprite.scale,s,v2(1.35,1.35))
+        const time=Numeric.clamp(1-(this.time/this.parachute_data.lifetime),0,1)
+        const s=v2(1.25,1.25)
+        v2m.scale(s,s,time)
+        v2m.add(this.sprite.scale,s,v2(1.3,1.3))
         this.sprite.position=this.position
+
+        this.sprite.zIndex=zIndexes.Airbodys+time
+        this.sprite.matrix=matrix4.parallax_2d(this.game.scene_2d.camera.position,1.1+(1*time))
     }
     override on_decode_net(stream:Stream,full:boolean):void{
         this.time=stream.read_float(0,30,2)