+++ changelog.md
@@ -19,7 +19,7 @@ ___
 * Hair Paints
 ___
 ### Others
-
+* Parachute Perspective
 ___
 ## Changes
 ___